apps/api/pyproject.toml
[tool.poetry]
name = "interview-beon-radpair-api"
version = "0.1.5"
description = ""
authors = []
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
fastapi = "^0.110.1"
httpx = "^0.27.0"
loguru = "^0.7.2"
python = "^3.12"
uvicorn = "^0.29.0"
python-multipart = "^0.0.16"
openai = "^1.53.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.0"
python-decouple = "^3.8"
google-generativeai = "^0.8.3"
gtts = "^2.5.3"
speechrecognition = "^3.11.0"
playsound3 = "^2.2.2"
pydub = "^0.25.1"

[tool.poetry.group.tests]
optional = true
[tool.poetry.group.tests.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"

[tool.poetry.group.ci]
optional = true
[tool.poetry.group.ci.dependencies]
ruff = "^0.3.5"
black = "^24.3.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
lint.select = ["I00", "E", "F", "W"]

[tool.black]
line-length = 120
target-version = ['py39']